Create SUM of SEquence
You are given a positive integer N
Your task is to make a recursive function and sum all the number until 
N reaches to 1 after performing the following two operations:-If N is even, then N= N/2 If N is odd, then N=3∗N+1
Repeat the above steps, until N becomes 1

Note:- It is guaranteed that after performing the above two operations N will become 1 at some point
Note:- Sum of all number can be large, so print the answer with modulo (10 ^9+7)
Note:-  - In real interviews, you have to use recursion when asked by the interviewer. So the coins will not be awarded if the solution is not recursive in the contest.

Input Format
The first line contains an integer T denoting the number of test cases. For each test case: an integer N is provided.
Output Format
For each test case print, the output in the new line: The sum all number until N reaches to 1 (including 1).