Cool
Mike loves solving Binary number problems. So he gives you a number N and asks you to find the binary representation of N
But he thinks this is easy for you, so he added one more condition to this problem. Now he asks to find all possible permutations of the binary representation of number N
Example:- 
Input: N=5,Binary representation of 5 = "101".
Output: "011", "101", "110".
Note:- Print the output in the lexicographically sorted form.

Note:-  - In real interviews, you have to use recursion when asked by the interviewer. So the coins will not be awarded if the solution is not recursive in the contest.

Input Format
The first line of input contains an integer 
T
 denoting the number of test cases.
For each test case, the input contains an integers 
N
.

Output Format
For each test case on a new line, print the permutation separated by space.